+++ net.tourbook.statistics/src/net/tourbook/statistics/StatisticMonth.java
@@ -79,8 +79,8 @@ public abstract class StatisticMonth extends YearStatistic {
 		 * create segments for each year
 		 */
 		final int monthCounter = tourMonthData.altitudeHigh[0].length;
-		final int segmentStart[] = new int[_numberOfYears];
-		final int segmentEnd[] = new int[_numberOfYears];
+		final long segmentStart[] = new long[_numberOfYears];
+		final long segmentEnd[] = new long[_numberOfYears];
 		final String[] segmentTitle = new String[_numberOfYears];
 
 		final int oldestYear = _currentYear - _numberOfYears + 1;
