+++ net.tourbook.statistics/src/net/tourbook/statistics/StatisticMonthHrZone.java
@@ -91,8 +91,8 @@ public class StatisticMonthHrZone extends YearStatistic {
 		 * create segments for each year
 		 */
 		final int monthCounter = monthData.hrZoneValues[0].length;
-		final int segmentStart[] = new int[_statNumberOfYears];
-		final int segmentEnd[] = new int[_statNumberOfYears];
+		final long segmentStart[] = new long[_statNumberOfYears];
+		final long segmentEnd[] = new long[_statNumberOfYears];
 		final String[] segmentTitle = new String[_statNumberOfYears];
 
 		final int oldestYear = _statYoungestYear - _statNumberOfYears + 1;
