+++ net.tourbook.statistics/src/net/tourbook/statistics/StatisticWeekHrZone.java
@@ -94,8 +94,8 @@ public class StatisticWeekHrZone extends YearStatistic {
 	 */
 	ChartSegments createChartSegments() {
 
-		final int segmentStart[] = new int[_statNumberOfYears];
-		final int segmentEnd[] = new int[_statNumberOfYears];
+		final long segmentStart[] = new long[_statNumberOfYears];
+		final long segmentEnd[] = new long[_statNumberOfYears];
 		final String[] segmentTitle = new String[_statNumberOfYears];
 
 		final int oldestYear = _statYoungestYear - _statNumberOfYears + 1;
